Performance Upgrades
Added in some JIT for performance speedups.

Also added in difference of sigmoid. One more idea to expand this is to use the "rights" parameters to add on to the values instead of using these values directly for the "b1" portion. Will need to check the performance of both of these methods before nailing something concrete